refactor

+++ setup.sh
@@ -20,7 +20,7 @@ test -L nvim || test -d nvim || ln -s ~/.vim nvim
 
 # for tmuxinator
 ln -s ${script_path}/tmuxinator                 # for mac
-ln -s ${script_path}/tmuxinator ~/.tmuxinator   # for linux
+test -L ~/.tmuxinator || test -d ~/.tmuxinator || ln -s ${script_path}/tmuxinator ~/.tmuxinator   # for linux
 
 # for bin/*
 test -d ~/bin || mkdir ~/bin
